refactor: moving resource group name into the azure folder
